Download data from the HCA via ExperimentHub

Download HDF5 (dense assay) and RDS (row and column annotations) files from the HCA via ExperimentHub, composing them together as a SingleCellExperiment object

HCAData(dataset = NULL, as.sparse = TRUE)

Arguments

dataset

A character string: which dataset should be retrieved?

as.sparse

Logical, specifies whether the underlying HDF5 dataset should be treated as sparse or not - will be passed to the call to `HDF5Array()`. Defaults to TRUE, i.e. by using the `DelayedArray` infrastructure.

Value

A SingleCellExperiment object with a HDF5Matrix in the countsassay, containing the UMI counts for each gene in each cell. Row- and column-level metadata are also provided.

Details

This current release includes the following datasets:

Census of Immune Cells - Umbilical cord blood

UMI counts from the 10x (droplet) single-cell RNA-seq data. The object contains counts for 33694 genes in 384000 cells.

Census of Immune Cells - Bone marrow

UMI counts from the 10x (droplet) single-cell RNA-seq data. The object contains counts for 33694 genes in 378000 cells.

Examples

HCAData()
#> No dataset name provided, please choose one of the following: 
#> ica_bone_marrow
#> ica_cord_blood
sce_cordblood <- HCAData("ica_cord_blood")
#> Retrieving Human Cell Atlas dataset: ica_cord_blood
#> snapshotDate(): 2022-03-01
#> see ?HCAData and browseVignettes('HCAData') for documentation
#> loading from cache